+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to `dash-ag-grid` will be documented in this file.
 This project adheres to [Semantic Versioning](https://semver.org/).
 Links "DE#nnn" prior to version 2.0 point to the Dash Enterprise closed-source Dash AG Grid repo
 
+## [31.0.1] - 2024-02-07
+
+- [#266](https://github.com/plotly/dash-ag-grid/pull/266) Updated README
+
 ## [31.0.0] - 2024-02-01
 
 ### Added